Update kalender kegiatan

+++ app/DataTables/UserDataTable.php
@@ -32,12 +32,20 @@ class UserDataTable
                     '</div>';    
                 }
 
+                if(Auth::user()->isAdmin()){
+                    return
+                        '<div class="btn-group">' .
+                            '<a href="' . route('user.edit', [$level, $user->id]) . '" class="btn  btn-sm  btn-warning" data-bs-toggle="tooltip" data-bs-placement="bottom" title="Edit" style="margin-right: 5px" ><b> Edit </b></a>' .
+                            '<a href="' . route('user.show', [$level, $user->id]) . '" class="btn  btn-sm btn-info" data-bs-toggle="tooltip" data-bs-placement="bottom" title="Lihat" style="margin-right: 5px" ><b> Lihat</b></a>' .
+                            '<a href="#" onclick="deleteModel(' . $deleteUrl . ',)" class="btn btn-sm btn-danger" data-bs-toggle="tooltip" data-bs-placement="bottom" title="Edit" style="margin-right: 5px"><b> Hapus</b></a>' .
+                        '</div>';
+                }
+
                 return
                     '<div class="btn-group">' .
-                        '<a href="' . route('user.edit', [$level, $user->id]) . '" class="btn  btn-sm  btn-warning" data-bs-toggle="tooltip" data-bs-placement="bottom" title="Edit" style="margin-right: 5px" ><b> Edit </b></a>' .
                         '<a href="' . route('user.show', [$level, $user->id]) . '" class="btn  btn-sm btn-info" data-bs-toggle="tooltip" data-bs-placement="bottom" title="Lihat" style="margin-right: 5px" ><b> Lihat</b></a>' .
-                        '<a href="#" onclick="deleteModel(' . $deleteUrl . ',)" class="btn btn-sm btn-danger" data-bs-toggle="tooltip" data-bs-placement="bottom" title="Edit" style="margin-right: 5px"><b> Hapus</b></a>' .
                     '</div>';
+
             })->addIndexColumn()->rawColumns(['action', 'foto'])->make(true);
     }
 }
